Move and clean up User model
